+++ CHANGELOG.md
@@ -6,6 +6,8 @@
  and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Removed `Copy to` option in the dropdown of the item in wish list.
 
 ## [0.1.3] - 2019-05-13